destructured Calendar.tsx, calendar navigation, highlights:
