+++ src/pages/Dashboard.tsx
@@ -1,11 +1,16 @@
-//import//
-
-import { Calendar } from "../components/Calendar"
+import { useEffect, useState } from "react";
+import { Calendar } from "../components/Calendar";
+import { TProject } from "../types";
+import { getDocuments, mapProjectDocument } from "../services";
 
 export const Dashboard = () => {
+  const [projects, setProjects] = useState<TProject[]>([]);
+  useEffect(() => {
+    getDocuments("projects", mapProjectDocument, setProjects);
+  }, []);
   return (
     <div>
-        <Calendar header="dashboard calendar"/>
+      {projects[0] && <Calendar header="dashboard calendar" model={projects} />}
     </div>
-  )
-}
+  );
+};
